refactor: step1 코드 리뷰 피드백 기반 리팩터링 1
