+++ src/main/java/mission/lotto/domain/Lotto.java
@@ -4,25 +4,58 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 public class Lotto implements Comparable<Lotto> {
 
-    private final LottoNumbers numbers;
+    public static final Integer LOTTO_PRICE = 1000;
+    public static final int MAX_NUMBERS_LENGTH = 6;
+    private final Set<Number> numbers;
 
     public Lotto(List<Integer> numbers) {
-        this.numbers = new LottoNumbers(numbers);
+        Set<Number> lottoNums = numbers.stream()
+                .map(Number::new)
+                .collect(Collectors.toSet());
+
+        if (checkNumbersLength(lottoNums)) {
+            throw new IllegalArgumentException("로또는 중복 없는 6자리만 가능합니다.");
+        }
+
+        this.numbers = lottoNums;
     }
 
     public Rank calculateRank(LottoAnswer answer) {
         Set<Number> answerNumbers = answer.getAnswerNumbers();
         Set<Number> combine = new HashSet<>(answerNumbers);
-        combine.addAll(this.numbers.getNumbers());
+        combine.addAll(numbers);
 
         int correctNo = 12 - combine.size();
         Number bonus = answer.getBonusNumber();
 
-        boolean hasBonusNo = numbers.isContainLottoNumber(bonus);
+        boolean hasBonusNo = isContainLottoNumber(bonus);
         return Rank.getRank(correctNo, hasBonusNo);
+    }
+
+    private boolean checkNumbersLength(Set<Number> temp) {
+        return temp.size() != MAX_NUMBERS_LENGTH;
+    }
+
+    public boolean isContainLottoNumber(Number lottoNumber) {
+        return numbers.contains(lottoNumber);
+    }
+
+    @Override
+    public int compareTo(Lotto lotto) {
+        return Integer.compare(Objects.hash(this), Objects.hash(lotto));
+    }
+
+    public Set<Number> getNumbers() {
+        return numbers;
+    }
+
+    @Override
+    public String toString() {
+        return this.numbers.toString();
     }
 
     @Override
@@ -36,15 +69,5 @@ public class Lotto implements Comparable<Lotto> {
     @Override
     public int hashCode() {
         return Objects.hash(numbers);
-    }
-
-    @Override
-    public int compareTo(Lotto lotto) {
-        return Integer.compare(Objects.hash(this), Objects.hash(lotto));
-    }
-
-    @Override
-    public String toString() {
-        return this.numbers.toString();
     }
 }
