+++ src/main/java/mission/lotto/domain/LottoAnswer.java
@@ -4,10 +4,10 @@ import java.util.Set;
 
 public class LottoAnswer {
 
-    private final LottoNumbers answerNumbers;
+    private final Lotto answerNumbers;
     private final Number bonusNumber;
 
-    public LottoAnswer(LottoNumbers answerNumbers, Number bonusNumber) {
+    public LottoAnswer(Lotto answerNumbers, Number bonusNumber) {
         if (answerNumbers.isContainLottoNumber(bonusNumber)) {
             throw new IllegalArgumentException("당첨 번호와 보너스 번호가 같을 수 없습니다.");
         }
