+++ src/main/java/mission/lotto/util/RandomUtil.java
@@ -1,21 +1,27 @@
 package mission.lotto.util;
 
+import mission.lotto.domain.Lotto;
+import mission.lotto.domain.NumGenerator;
+import mission.lotto.domain.Number;
+
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-public class RandomUtil {
+public class RandomUtil implements NumGenerator {
 
-    private static final int MIN_NUMBER = 1;
-    private static final int MAX_NUMBER = 45;
-    private static final List<Integer> numberList = IntStream.range(MIN_NUMBER, MAX_NUMBER + 1)
-                                                             .boxed()
-                                                             .collect(Collectors.toList());
+    private static final List<Integer> numberList;
 
-    public static List<Integer> getRandomSixIntegerList() {
-        Collections.shuffle(numberList);
-        return numberList.subList(0, 6);
+    static {
+        numberList = IntStream.range(Number.MIN_NUMBER_VALUE, Number.MAX_NUMBER_VALUE + 1)
+                .boxed()
+                .collect(Collectors.toList());
     }
 
+    @Override
+    public List<Integer> getSixNumbers() {
+        Collections.shuffle(numberList);
+        return numberList.subList(0, Lotto.MAX_NUMBERS_LENGTH);
+    }
 }
