+++ src/test/java/mission/lotto/domain/LottoAnswerTest.java
@@ -13,7 +13,7 @@ public class LottoAnswerTest {
     @DisplayName("LottoAnswer 생성 예외 테스트 ")
     public void 정답계산() {
         assertThatThrownBy(() -> new LottoAnswer(
-                new LottoNumbers(Arrays.asList(1, 2, 3, 4, 5, 6)),
+                new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6)),
                 new Number(5)
         )).isInstanceOf(IllegalArgumentException.class);
     }
